Correctly emit error events and return error on pathfind
WIP:
TODO: Catch rippled errors in RippleAPI and throw a RippleError. Include the
original rippled error as a property on the RippleError ()

Request#complete used a function to determine success/failure of a pathfind
request. The "filterFn" was just a Boolean. If a response from rippled came
back, it considered that a success, even for a response with an error.

Additionally, if a pathfind request errors, the request does not be closed,
which leads to this error:

    {
        "error": "noPathRequest",
        "error_code": 31,
        "error_message": "No pathfinding request in progress.",
        "id": 2,
        "request": {
            "command": "path_find",
            "id": 2,
            "subcommand": "close"
        },
        "status": "error",
        "type": "response"
    }